refactor: change api prefix
