+++ server/src/api/auth/index.ts
@@ -1,0 +1,3 @@
+export { default as authRouter } from './routes';
+
+export default {};
