+++ server/src/api/index.ts
@@ -1,12 +1,12 @@
 import Router from 'koa-router';
-import { userRouter } from '@api/user';
 import testRouter from '@api/test/test.routes';
+import { authRouter } from '@api/auth';
 
 const rootRouter = new Router({ prefix: '/api' });
 
 rootRouter.get('/', () => ({ test: true }));
 
-rootRouter.use(userRouter.routes());
+rootRouter.use(authRouter.routes());
 rootRouter.use(testRouter.routes());
 
 export default rootRouter;
