--- server/src/api/user/index.ts
@@ -1,3 +1,0 @@
-export { userRouter } from './routes';
-
-export default {};
